refactor: CoreDataStack.swift 파일 생성 및 내부 구현
