+++ CloudNotes/CloudNotes/CoreDataStack.swift
@@ -1,0 +1,34 @@
+import CoreData
+import Foundation
+
+class CoreDataStack {
+    private let modelName: String
+    init(modelName: String) {
+        self.modelName = modelName
+    }
+    lazy var persistentContainer: NSPersistentCloudKitContainer = {
+        let container = NSPersistentCloudKitContainer(name: modelName)
+        container.loadPersistentStores(completionHandler: { (storeDescription, error) in
+            if let error = error as NSError? {
+                fatalError("Unresolved error \(error), \(error.userInfo)")
+            }
+        })
+        return container
+    }()
+
+    func saveContext () {
+        let context = persistentContainer.viewContext
+        if context.hasChanges {
+            do {
+                try context.save()
+            } catch {
+                let nserror = error as NSError
+                fatalError("Unresolved error \(nserror), \(nserror.userInfo)")
+            }
+        }
+    }
+    
+    func showErrorAlert() {
+        
+    }
+}
